Show warning for colliding BA2 hashes
BA2s will override each other when the hashes of their file names
collide. This will cause the later-loaded BA2 to completely overwrite
the previous BA2.
The only way to resolve this is to rename one or both of the colliding
BA2s. This should be done by mod authors so that all users can benefit,
obviously.

This commit adds a warning for these cases. Fallout 4 only, since we
don't know if the other games have such hash collisions too.

Closes #558
Ref #542 - we need a proper warnings API