Path 'name' param renames WIP:
Proved a rabbit hole - main result is a bit faster and hopefully cleaner
code:

This' parent _main including installers

    18148523 function calls (18121204 primitive calls) in 12.068 seconds

This

    18111904 function calls (18084585 primitive calls) in 11.571 seconds

EDIT: don't pay much attention to __getCase will be removed in next
commit. I kept it separate to see what changes in the tests but better
viewed together.
getCase and getNorm are slightly different - added types to
point the diffs out. getCase is used in comparisons when the Path case
is ruled out. getNorm is used to convert str_or_path, but not None to
unicode - it would blow in all its three uses if None was passed.

In getNorm note return u'' # and not maybe b''

It seems like init would not normalize the path always - decided to
push the API a bit and only accept normalized paths (push in the sense
not enforce). Currently this seems to be the case anyway.

Squash: test current behavior

1 test == 1024 docs although we need 1024 tests

decoder -> move the encoding u'cp65001' check when needed.

Under #368