Begin reworking Tweak Settings
This lays the groundwork for getting rid of the GmstTweaks/GlobalsTweaks
constants and turning them into classes instead.
This is more readable, allows the IDE to reason about them and will
allow us to deduplicate settings that are shared by multiple games.

Differences in tweak_choices between games are no problem - just union
them - but differences in defaults in those choices are a problem. I
introduced a defaultdict to handle this, but it's a bit of an ugly
solution.

This commit currently only implements the World: Timescale tweak for all
games, as a POC.

Mopy/bash/patcher/base.py:
Tiny performance/elegance improvement