Always keep patcher and dialog checkboxes in sync
Before this commit, if you used certain buttons in the patcher dialog,
the checkboxes in the GUI patcher sources on the right and the ones in
the patcher list on the left would stay in sync. For example, using the
Select/Deselect All buttons would do this. Furthermore, they would also
enable/disable the Build Patch button as appropriate.

However, this did not extend to all other ways of enabling/disabling
patcher sources. For example, if you first used the global Deselect All,
all patchers and all sources got deselected and the Build Patch button
disabled. However, if you then activated one source in, say, the Merge
Patches patcher, the patcher itself did not get checked and the Build
Patch button did not get activated. You had to manually check the Merge
Patches checkbox for that to happen.

As of this commit, that has been fixed. Checking sources will enable
the patcher, unchecking all of its sources will disable it again. The
Build Patch button will also always be updated correctly.

Simplify OnListCheck

AFAICT, this is equivalent (though the logic is somewhat complex):

`lb_selection_dex in range(len(self.items)`) always holds. Therefore,
new_checked is always False (because any_checked must be False for us
to get here) and this entire branch can be replaced with
`self._enable_self(False)`. At which point we can replace the entire if
check with self._enable_self(any_checked).