Hunt down 'Data' usages in strings
Need to be replaced with bush.game.mods_dir, since it's called 'Data
Files' for Morrowind (plus will help us out for later games). Left
'Data' occurences in link names alone (e.g. 'Clean Data') so that we
don't have to special-case that part of our docs.

Also contains a rewrite of the Anneal/Anneal All help text, which seems
to be one of the most confusing bits of BAIN.

Under #479