bain: Move package after Sync From Data
If trying to sync a RAR from Data, we have to create a new 7z file. It's
annoying that it always lands at the bottom of the installer order and
then has to be installed again to get it to be active, even though in
most cases we can know the contents are going to be almost the same.