Equalize performance of {dis,en}abling plugins
Disabling took some 16000 more function calls and was therefore a bit
slower.

Cache dependent plugins

Cleans up Mod_ListDependent as well. But the main attraction is of
course the ~6000 function calls that have disappeared when disabling a
plugin. Still doesn't bring the two to equal speed, but it's an
improvement.

Avoid get_ordered in lo_deactivate

Turns out sorting the entire load order isn't that cheap. This chops off
10000 calls from disabling, bringing enabling and disabling pretty much
in line.