Pass redraw to RefreshUI, round 1
The big one. Kept this for close to last because I knew it would hit
hard, especially on the time saved. And it does, with ~150000 calls and
~0.22s gone.

We're not entirely done yet, disabling still takes ~0.02s longer than
enabling, which just barely pushes it over the 0.1s threshold.

Refresh higher-loading plugins too

UGH. Forgot about the stupid columns. We can't get away with just
UI-refreshing the plugin and its masters/dependents, but actually have
to UI-refresh all higher-loading plugins than the lowest-loading plugin
that was involved in the LO operation.

We really only need to change one tiny bit of the UI for all these extra
plugins (one or two columns), but there is no way for us to communicate
that fact to RefreshUI. So with this commit the GUI is always updated
correctly and it will be just as fast as before if you (un)check among
the last couple plugins in the load order, but unchecking one of the
first couple plugins brings the old 0.3s back :(

Refresh imported and merged plugins too

Forgot about bashed patches...
Note that we don't need to do this with imported/merged mods that are
active, since their checkboxes won't change (and we only need to do this
refresh in the first place because of the checkbox icons - again, would
be lovely if we could tell RefreshUI that :P).