Pass redraw to RefreshUI, round 3
This finishes off the branch by applying the same optimization to drag
and drop as well as Ctrl+Arrow keys.

Here we are working directly with indices, so the logic is very simple -
just find the smallest index and pass that to _refreshOnDrop, which can
then look up the impacted plugin and pass it to _lo_redraw_targets.