Fix flicker due to global menu
Finally figured out what it was. As per usual, there are two common
ways to cure flicker:
 1. Skip wx calls when they're changing a property to the same value it
    had before the call, thereby causing tons of unnecessary redraws.
 2. Freeze before doing tons of changes, then Thaw afterwards.

In this case, it was no. 2. Took the opportunity to de-wx Freeze and
Thaw while also wrapping them in a more pythonic API.

Use DestroyItem instead of Remove

Changes nothing, but saves a line or two.

Under #190