Comb over all StringIO usages
Replace them with io.String or io.BytesIO depending on usage. Ugh, this
was really tedious and much harder than expected.

This commit is probably best viewed with whitespace diff off, because I
took the opportunity to get rid of bolt.sio and all close() invocations
(StringIO/BytesIO never need to be closed since they don't point to OS
resources - the garbage collector will do a much better job than we
could). Another benefit of that is getting rid of one indentation level,
which unfortunately makes this commit impossible to read with whitespace
diff on.

Note the usage in converters.py - no idea if that works. AFAICT it
should, even on py3, but we may want to instead pass the new encoding
kwarg to Popen once we are on py3 to open the streams in text mode
instead.

Mopy/bash/bosh/_saves.py:
Some prefixing etc. in already touched lines and ones in the immediate
vicinity