Upgrade settings to VDATA3
We use UTF-8 as the guess, since it's backwards compatible with ASCII.
The settings should be entirely ASCII, BAIN data should only be in
bytestrings if it's *really* old (like, pre-unicode-merge old).

We make a permanent backup copy before saving the converted settings as
'Table-vdata2.dat.bak' (with 'Table' replaced with 'BashSettings',
'BashLoadOrders', etc. as appropriate).

Under #460
Closes #597