Get rid of all hardcoded color assignments
Add a new color entry to customize the warning color,
and use the default color when no warning exists

Tested with both the default windows theme and a dark theme.

Infernio: This is here because it also affects Linux, fixing the default
color and allowing you to customize it otherwise.

Allow new color settings without error

Previously, the color settings would raise a key error on
the first start of Wrye Bash when a new color setting key
was aded, because the default settings for it was only
copied if it was missing, not changed.

Infernio: Reworked this to be a generic 'warning' color instead and used
it to replace all hardcoded usages of colors throughout WB :)

Get window background color from system

Co-authored-by: Infernio <infernio@icloud.com>