Create a custom wizard API
Completely de-wx'd. Only problem is that it has a tiny bit of flicker
when changing pages. I can't tell what the wxWidgets wizard
implementation does differently that lets it avoid that. In the end, I
decided not to let perfect be the enemy of good. It's a tiny bit of
flicker, which is a small price to pay for fixing the horrible back
button issue and getting entirely de-wx'd wizards.