[!]MelBase makes default -> set_default a kward-only arg:
Also minSize -> is that used anywhere? (!)
Other than that is refactoring to make defaults handling more explicit.
I initially wanted to remove the 'default' argument altogether, but
we might need to specify defaults and add a with_defaults(dict) to
MelSet would duplicate the attr names. Plus we might want to make some
attrs required, as the code seems to suggest, so the new set_default
was put to use in skyrim/records - should we? (!)

Note I edited MelFixedString(MelStruct) - it should be really a MelBase
(rather MelAttr to be) subclass. The confusion extended to the point of
passing it bytes defaults.