Rework active plugin cleaning
Move all of it out of AsteriskGame, TextfileGame needs to be able to do
the same thing with the game master and Update.esm. The game master was
actually already being handled, but in a very ad hoc fashion. Now it's
handled through the same system that also handles the DLCs in SSE.

Basically, here's what needs to be removed from actives:
 - Morrowind, Oblivion: nothing
 - FO3, FNV: nothing (haven't verified this though)
 - TES5, Enderal: Skyrim.esm & Update.esm
 - SSE, TES5VR, FO4, FO4VR: game master, DLCs, CC content

Under #578