Flush deprints to BashBugDump instantly
Right now, we're in the unfortunate position of writing enough data to
the BashBugDump that we often flush it in the middle of a traceback,
making people think that's all there is and just uploading that.
Instead, we can open the BashBugDump in unbuffered mode and simply flush
it on every write.

Mopy/bash/bash.py:
Fix Tkinter error dialog not exiting

Yep, sys.exit should be in there too.