Restrict 'Disable' command to older games
This is a terrible idea for newer games - removing a plugin from an
existing save can brick that save nicely. It is apparently quite safe
in older games (i.e. pre-Skyrim) though.

Since we're keeping it around, I fixed the implemention to be less hacky
and to disable it for non-ESM files.