fomod: Fix regression when installSteps lack order
Technically a regression in that it worked way back when Ganda first
added FOMOD support. I broke it during some subsequent refactoring and
no one noticed because pretty much no mod author is stupid enough to not
put an order="Explicit" on their installSteps. Sure, let the mod manager
figure out the order in which to show the wizard's pages...