cosaves: Only handle DATA chunk for SOS
Turns out DATA is a very common signature (go figure) so we need to only
handle it when we know it's a DATA chunk from SOS.

Rest is some repr's that helped a lot in debugging this. One of them is
very hacky, but since it's just a repr it's okay for now.