Block bak enabling, allow quick/autosave disabling
Enabling/disabling a quick/autosave makes sense, but it was probably
blocked because of the backups that quick/autosaves can have. Now that
we read arbitrary bak files on the Saves tab, we really just need to
disallow enabling all backup files (with an error message of course) and
can allow disabling and enabling of auto/quicksaves.

Under #580