Fix regression in SEWorlds Patcher
Didn't get updated for our CTDA rework all the way back in
f0147827e0480797f3f59f723bd464fa237b296d.

We can't just rely on defaults here since each function has a different
format. In this case, we have no parameters and a float comparison
value, so we need to set those types for param1, param2 and compValue
(right now, param2 will already be set to that, but that relies on the
return value of next(itervalues), which is not guaranteed to remain the
same across systems/python versions/etc, so we have to set param2 as
well to be sound).