Split out race tweaks to new file
First step to splitting this nightmare of a patcher. The tweaks are
pretty well contained, ~~so they're probably the easiest part to
untangle.~~ Hello, me from the future here. That's a lie, they're
actually the only thing that didn't just work™ once split.

RaceTweaker -> RaceTweak

To match the other tweak class names.