Fix race patcher forwarding skills incorrectly
Well, it's not *technically* incorrect, but this was definitely not what
anyone was expecting. Basically, when more than one mod touches the
record, it would 'merge' the skills because the skills are sorted. That
means you'd end up with skill boosts on whatever skill the next mod in
the load order had forwarded at that point. So e.g. if I had a mod that
changed a race's 'Blades' skill to 'Speechcraft' and boosted the
'Unarmed' skill from 5 to 10 and another mod after it that leaves the
skills unaltered, I could end up with the Speechcraft change getting
forwarded, while the skill boost was instead applied to the 'Heavy
Armor' skill because the order of skills is different between the two
mods.

Instead, we now store skills and skill boosts as a single list and
forward that entire thing. This fixes it, but the implementation is
nasty - normally we'd use MelLists for this, but this particular
subrecord has flags, which MelLists doesn't support. So I had to hack
together something to manage this for now. Proper solution is
refactoring MelLists to support flags.