FO3, FNV: Port Import Races
Also correct a bunch of incorrect docs for the R. series of tags.
There's no difference between FO3's and FNV's RACE, so this works for
FNV immediately as well.

Add aliases for Eyes-D, Eyes-E and Eyes-R

Found them again in valda's codebase, where they're used interchangeably
with the 'Eyes' tag, which is now 'R.Eyes' in our codebase.