Add Reorder Masters command to Saves tab
Requested on Discord, but this has been something I wanted to add for
quite a while now. The implementation of lo_reorder isn't the prettiest.
Ideally we'd use a topological sort here, but I didn't feel like
writing graph boilerplate or pulling in a library, so this will have to
do. Pretty fast in practice.

Also reworked the Saves context menu readme section.