Drop _process_csv_sources:
A joy forever. Note that _called_from_patcher is also used in
FactionRelations that do not seem to be used by the patcher - only four
parsers appear as _csv_parsers:

    _csv_parser = parsers.ActorFactions
    _csv_parser = parsers.FullNames
    _csv_parser = parsers.SpellRecords
    _csv_parser = parsers.ItemStats

We could add those in a mixin with an init override to further isolate
_called_from_patcher but it's rather noisy - plus, as noted, the model
should be moved to records - this will solve most such problems at once.

Under #312