Split fsNames into different vars
Nice side effect is being able to drop the FNV template since it's
identical to FO3 and the only reason it was seperate is because we used
fsName for this check.

Also added a bunch of deprints for detected paths to aid in debugging
user reports.

Support windows store games AppData locations

Required reworking the game detection to return a bit more
information, and splitting game.fsName into two attributes:
`fsName` - used for the game location and My Games saves folder
`appdata_name` - used for the AppData folder

Update saves locations for WS games

Co-authored-by: lojack5 <1458329+lojack5@users.noreply.github.com>