is_archive/project erosion:
Note I replaced the io buffer with a plain list and join - I find this
more explicit and should be faster too (I mean io probably does this
internally so we are minus the function calls). Similar for concatenating
strings - generally the pattern is join - has the advantage of cleaning
the \n from the strings making them more readable.

More work on _dumpFiles:

- fstrings - note exception (would not wrap)
- initially I defined a

    markstr = lambda x: ' N/A' if is_mark else x

but hey don't calculate the string if not necessary

Looks like a helper class for info pages is needed with a
_dump_files and a dict for holding the info pages singletons, key'd
by pageName -  or not

squash! is_archive/project erosion:

Cheated a bit using _valid_exts_re to tell archives apart - actually
this is more correct, we do care here about the Installer having a
meaningful extension rather than it being an archive.

strings_exts was unused