Bold new patchers that don't have items too
The docs and help text claim that 'Items that are new since the last
time this patch was built are displayed in bold', but that doesn't
actually apply to anything but new patchers with items automatically
selected. Which is probably the majority of cases, but it's still worth
making this work for the future.

Under #587