docs: Remove Template Flags from Actors.ACBS
Template Flags are not supported by Wrye Bash yet, because no one has
reverse-engineered what exactly each of the flags applies to.
Furthermore, similar to the Partial Form flag (#577), we would
need much more refactoring before we could make the BP handle these
correctly (it would have to merge the previous version of the record
with the new one if a template flag is present, based on the flags that
are applied). Instead, remove them from the documentation for now.

Under #588

Contributing.md:
Update link to [dev] Commit guidelines article

The link to & name of the article has changed.