Disable xSE, LAA and ScriptDragon for WS games
Even if the user installs these, they won't work, so we don't want
to provide UI for them if they exist.

Refactors out some common boilerplate when defining a WS game,
but could be a little too much "black magic" for our tastes,
specifically the `game_detect_includes` override.

Infernio: Note that LAA is still unused, we should either implement or
drop it.