Set allTags dynamically during startup
Eliminates one of the most annoying places where conflicts arise. We
already store most of the tags in the patchers, so this is duplicate
information as well.

I don't love that this has to modify a game constant, but we don't
really have a great place for such a constant yet.