Supercharge the performance of 'Details'
Well actually, it's slightly slower (~20%) than before, but that's
because the previous implementation skipped 'complex' groups (DIAL,
CELL, WRLD) entirely. Those happen to be the biggest *by far*, so the
fact that we're almost as fast while processing an order of magnitude
more data is incredible.

Actually introduced this because I want to rework quite a bit of the Mod
Checker. We should be able to do once pass over the plugins and use the
collected info to display a lot of useful info about plugins, not just
whether or not they have deleted references.