Thin balt further, rework List Mods
The 'hold down C to show CRCs' doesn't work at all because holding down
C of course jumps to the 'Columns' menu entry. That's entirely
reasonable and desired behavior and I have no intention of breaking it.
Instead I changed this command to use Shift instead (and added the
ability to hold down Ctrl to skip the versions in the output).

The rest is balt erosion, as the title implies.

Under #190