Warn about HITMEs
HITMEs are records that have a higher mod index (i.e. the first byte of
the FormID) than the number of masters of the containing plugin. In
other words, if you imagine a file foo.esp with two masters, Skyrim.esm
and Update.esm, the situation is as follows:

 - 00000800: Override or injected record into Skyrim.esm
 - 01000800: Override or injected record into Update.esm
 - 02000800: New record
 - 03000800: HITME
 - 04000800: HITME
 - etc.

Mopy/bash/bolt.py:
Slight docstring fixup