Sort records before dumping
Those need to be sorted by FormID (or in the case of DIAL, the PNAM
subrecord). We were already doing that for CELLs (getBsbSizes does the
sorting, added a comment there clarifying what it's doing), but the
other groups need to sort as well - otherwise the game engine could end
up not finding a record it's looking for.

Under #480