Add a framework for showing BP config errors
Simply raise a BPConfigError at any step of the BP process (with a
translatable error message) and it will be shown to the user with a
notice that the configuration is incorrect.

Also drops the [1:] from TweakNamesPatcher.__init__, which seems to be a
regression from 120234c8fe6be6fa9698ce26ef5e553349e71745.