FO3, FNV: Port Tweak Names
Brings all the tweaks present in valda's version into our version. With
this, we are finally at feature parity.
Nice side effect is that you can now choose a custom format for each of
the tweaks. This was impossible to implement pre-tweak-pooling because
we lacked a way to validate that the format will work, but TP gave us
validate_values.

Closes #468

Co-authored-by: valda <valda@underscore.jp>