Convert/Remove newlines from details fields
Author, player name, etc. will be displayed in TextFields. New lines in
those fields can appear anywhere on the string (hence we can't use just
strip()) - their origin is unknown but are encountered in the wild.
wxPython really doesn't like a newline in such a field (understandably)
and flickers like crazy when one does appear. Description is a TextArea,
so newlines are fine, but wxPython will automatically convert them to
UNIX-style newlines when non-UNIX newlines are passed in. That results
in a bogus callback that enables the Save/Cancel buttons for no reason.
Instead, replace the newlines with UNIX-style ones at the
brec/save_headers API boundary and remove them for the single-line
fields.