Don't dump XACT flags if equal to default
This subrecord blocks activation when it's present without any flags
set. This isn't the right fix, that would be changing the default to
None, but there are way too many uses of flags without checking if
they exist in the codebase (plus all those flags are called 'flags',
which really doesn't help) to do that right now. Plus we'd need the
ability to mark subrecords as required for the couple flags that really
*do* have to be present in all cases (e.g. QSDT).

Mopy/bash/patcher/patchers/checkers.py:
Small performance improvements in Contents Checker