Skip vanilla deleted references in Plugin Checker
Rely on LOOT for those instead. Reason is that some of the vanilla files
in Oblivion have deleted injected records, which QAC can't clean (duh).
We *could* detect those, but it's a lot of effort: for every deleted
FormID we would have to build a load order FormID, then look up all
versions of the FormID and check the number of versions and check if the
first version's origin plugin matches the one the FormID resolves to.

Rather than bother with that, let's just rely on LOOT for the cleaning
report.