Add disclaimer to unlimited jewelry tweaks' tips
LOOT has a message warning users not to use this tweak with an old
version of Hemingwey's Capes, but the issue applies to all versions.
Since many of the capes use the Amulet Biped Flag, the tweak will apply
to them as well, causing them to not display. However, I think a more
appropriate solution would be to inform users in the tweak's tooltip.
That way, if they still want to use it, they'll know it can apply to
more than just literal amulets & why. Not all of the capes are purely
aesthetic, so some users may still want to use the tweak with this mod.
Since the same logic also applies to rings, I added a disclaimer to
that tweak too.

Infernio: Moved the Hemingwey's Capes note to the readmes. And because
vim on Windows was stupid enough to put tabs into those edits, I took
this opportunity to rid our readmes of all tabs.