Replace reTesNexus regex with new one
No idea what the old one was doing. What was the point of those bain and
tessource checks in there? Plus it was broken, e.g. searching for USSEP
on Google would search for 'Unofficial Skyrim Special Edition Patch 266
4 2 4b 1608419879'. It was supposed to just be 'Unofficial Skyrim
Special Edition Patch', which this new one accomplishes. And on top of
all that, it's also more than three times as fast.