tests: Add tests for LOOT condition parsing
Had a serious bug in my initial not-expression implementation (see
TestCanonicalNotExprComplex), so it's time for some tests.

Needs some changes to loot_parser.py/loot_conditions.py to make __repr__
always put out a stable, canonical representation that can also be
parsed if fed back into the parser.

Threw in a new exception class to get rid of the ParserError usage in
loot_conditions as well - those are definitely not parsing errors. Also
made _process_path error on paths outside the game folder. That's an
undocumented behavior of libloot, but it's sensible.