Fix regression in save face editing
Broke in 4d031ef55bc08003ff1f33df50ca4c0d245db281, which renamed 'level'
-> 'level_offset' in NPC records. I could have reverted that, but I
decided to simply take it further and rename the usages of 'level' in
saves/faces code as well.

Also contains a followup fix for
0164fff4bd35a799b33a5f6bfe347388ef638c85, which fixed a regression only
when reading saves but forgot to fix that same regression for writing
saves.