loot_conditions: Implement short-circuit evaluation
any(), all() and len() all have to evaluate the regex matches for every
file in the wanted directory, which is usually the Data folder and so
has tons of files in it. By instead just looping manually, we can abort
as soon as we hit the conditions that allow us to short-circuit. It
might even be faster in the worst case because we skip the method call
overhead for the call into any/all/len.

Also fixes _fn_many misusing the Path API. Never reported because no
LOOT tags use a many() condition yet.