Drop rp_exists:
For a more EAFP experience. Turns out the two methods were exactly
parallel, differing in one issuing hasattr while the other getattr.
The performance should be very similar (hasattr uses getattr internally)
and in the cases that rp_exists returned true we would actually
duplicate the resolving of the path. Those cases would be more numerous
than it might seem as we setDefault on record elements, so hasattr would
seldom fail. Still it took me a while to grok, precisely because the
semantics of get/hasattr for records are blurry. Actually this was
triggered by hasattr changing semantics in 480-pt3, returning True for
all valid record attributes on a wip attempt to drop setDefault,
so this code completely breaks - anyway the simplification of the
RecPath API is worth it :)

Under #480