Refactor Flags:
This evolved in several steps - first I dropped the value parameter from
init then I made getNames to a from_names factory, then introduced
TrimmedFlags - the initial motivation was to get rid of the
Flags(0, getNames) let's say peculiar syntax and to investigate if we
can switch to the IntFlag built in Enum. This does not seem trivial,
plus this refactoring gets us to an acceptable level of pythonism.
Instead of creating an instance of Flags and then use __call__ to
copy the input data we create an subtype of flags we use to create
those instances. Flags lost two instance attributes that are now
class variables - the second one became eventually a subtype. All this
is good old OOP - explicit is better than implicit.

The __call__ override would be nice to go but it's not terrible plus its
implicitness makes it hard to uproot - let it be for now. Anyway the API
of those Flags does not parallel the IntFlag interface so they might as
well remain as an idiomatic bolt API for a while.

Notes:
- Mopy/bash/brec/utils_constants.py: BipedFlags I dropped flag_default,
was unused.
- Mopy/bash/game/skyrim/records.py: note _RaceDataFlags1, I hadn't
- Mopy/bash/game/fallout3/records.py: dropped DecalDataFlags was unused

Remove pickling from bolt.Flags

I don't think we pickle those and I certainly don't want to pickle them.
Also some deduplication in records (aiService flags)

Formatting

Note that I chop off the final coma and that I avoid specifying
the index when there are no gaps, it does not really carry any
information (especially when there are only a couple names) -> edit:
let's say up to three values - the indexes are not what the flags are,
they are 2**index, but for bigger Flags it's easier to see the index
than count.

WIP regex for multiline ones:

https://intellij-support.jetbrains.com/hc/en-us/community/posts/206373939-Possible-to-make-match-line-breaks-Replace-in-path-?page=1#community_comment_207462899

\(Flags\.getNames\(([\s\S]*?)\)\)

.from_names\($1\)

With lots and lots of manual tweaking

Use a specialized subclass for "unknown_is_unused":

This removes all the processing from the main Flags class and aligns
better with IntFlag, however I am not sure that Flags is-a IntFlag so
this is shelved for now.

Took the opportunity to add slots to subclasses missing it.
Added __copy__ just in case - we should investigate if the memo dict is
correctly used in __deepcopy_

I dropped the sorting in getTrueAttrs - self.__class__._names should be
sorted in values by construction

XXX hunt down __call__

I gave up in Mopy/bash/brec/common_records.py (note the Flags copy!
comment) - note also _flag_default in Mopy/bash/brec/basic_elements.py
and

@@ -1119,3 +1116,3 @@ def __and__(self,other):
         if isinstance(other,Flags): other = other._field
-        return self(self._field & other)
+        return self.__class__(self._field & other)

Never use the constructor - we need the class Factory:

@@ -101,3 +101,3 @@ class MelBipedObjectData(MelStruct):
     """Handler for BODT/BOD2 subrecords.  Reads both types, writes only BOD2"""
-    _bp_flags = BipedFlags()
+    _bp_flags = BipedFlags.from_names()

Under #480