docs: Add WeaponMods
Also rename the patcher to match our naming scheme, fix a typo and
remove patching of completely unrelated attributes from WeaponMods
(scopeModel and hasScope flag).

Note that WeaponMods has some overlap with Graphics and Sound, in which
case WeaponMods is (obviously) preferred and the functionality of the
other tags is only retained for backwards compatibility.