Switch to FOMOD mode sooner when running wizard
Wizards need data from Installer.espmMap, which is populated in
refreshDataSizeCrc. In order for it to do that, however, the installer
(and GUI) cannot be in FOMOD mode when refreshing. Otherwise leads to
nonsense wizard errors being reported when running a wizard in FOMOD
mode:

belt.pyo  750 Continue: Error in wizard script
Traceback (most recent call last):
  File "bash\belt.pyo", line 748, in Continue
  File "bash\ScriptParser.pyo", line 547, in RunLine
  File "bash\ScriptParser.pyo", line 411, in __call__
  File "bash\ScriptParser.pyo", line 312, in __call__
  File "bash\ScriptParser.pyo", line 272, in execute
  File "bash\belt.pyo", line 1409, in kwd_de_select_plugin
  File "bash\belt.pyo", line 1417, in _select_plugin
  File "bash\ScriptParser.pyo", line 219, in error
ParserError: (Line 319, Column 70): Plugin 'Better Cities IC Waterfront FPS Patch.esp' is not a part of the installer.