543 - Stop comparing bytes and str subtypes with Path:
Triggered by FName comparisons with Path being non symmetrical. We have
to disallow Path comparisons with str subtypes and also with bytes to
further isolate Paths and more closely follow the py3 type idioms.
We could also patch GPath but that would be a (performance) overkill -
we don't seem to pass bytes in ever - cf __setstate__, which does call
decoder - that's ok. The decoder calls in the comparisons are welcome to
go though.

543 radiation nit:

Mopy/bash/gui/popups.py: no need to GPath here - defaultDir is a Path
always and defaultFile usually a str (or an FName to be) which needs
no GPathing just to get it back using .s. I followed usages chopping
off u' - not a good idea still (conflicts)

- 'deleted' renames
- patcher_dialog: unused import

Infernio: don't put two statements on one line

Just randomly noticed the one in bosh/__init__.py and did a search for
any in-code semicolons. Not exactly 543-related, but hey.

Under #368, #543.