Couple more random mac fixups:
Edit control is None on mac when we click on an installer name

If long pressing this triggers the rename although F2 handler is not
registered but when I was short pressing would crash trying to access
attribute SetSelection on None.
ec_set_selection should not return anything

Mopy/bash/gui/base_components.py:
reset_background_color() crashes on mac

clearReadOnly WIP:

https://superuser.com/a/40754/35020 for darwin - I used the linux variant
finally. Should be unified with clearRO and passed as an onerror callback
if possible