Fix persistent CELL handling in Import Cells
It was copying the persistent CELL block over wholesale (including all
child records) just because we wanted to import part of the CELL. This
shouldn't cause actual issues (in theory) because our FormID trimming
(using keepIds and mergeIds) would just discard these at the end, but it
is useless copying that we can easily avoid by doing it properly.

The other changes are a minor performance improvement by avoiding the
constant '.cell' accesses in handlePatchCellBlock.