compiled: Bump 7z to 21.05
No security fixes as far as I can tell, but plenty of bug fixes and some
performance enhancements - especially for people with more modern CPUs.

Doing this on dev because GitHub does not clean up orphan commits unless
you specifically ask them to, so when we rebase these would get stuck in
our history. Did some testing and everything works as expected, output
format does not seem to have changed in any way (and the changelog
doesn't list anything of the sort either).