Fixup for BP dialog not being destroyed:
Well this is weird - on mac when clicking the X button on a modal dialog
Close evt is fired ok and on_closing handler is triggered - then the UI
hangs. Clicking on the UI makes the mac-error-sound and on the debugger
the _post method of the activate event is shown in the frames window,
*below the gui.top_level_windows.DialogWindow.show_modal* line:

    return self._native_widget.ShowModal() in (_wx.ID_OK, _wx.ID_YES)

So it seems the __exit__ is never called for those and we eternally stay
in the `with (...) as dialog:` statement.

Probably this is the issue with windows user who need to escape/enter
the invisible BP build dialog:
https://discord.com/channels/537706885965676554/537710082755133460/913812302975873066

Also (some of it is outdated but still):

https://stackoverflow.com/a/21391171/281545
https://stackoverflow.com/a/34300474/281545
https://stackoverflow.com/a/19815382/281545

Mopy/bash/basher/mod_links.py: master errors -> use show_modal
We should encapsulate show_modal inside display_dialog as much as
possible - see next commit.

Note this might be the root cause of:

Traceback (most recent call last):
  File "bash\gui\events.py", line 177, in _post
  File "bash\basher\settings_dialog.py", line 130, in _send_apply
  File "bash\basher\__init__.py", line 3938, in Restart
  File "bash\gui\top_level_windows.py", line 100, in close_win
wx._core.wxAssertionError: C++ assertion "nNew != dynamicEvents.size()" failed at ..\..\src\common\event.cpp(1917) in wxEvtHandler::SearchDynamicEventTable():
Traceback (most recent call last):
  File "bash\balt.py", line 1770, in __Execute
  File "bash\basher\misc_links.py", line 358, in Execute
  File "bash\gui\top_level_windows.py", line 187, in display_dialog
  File "bash\gui\top_level_windows.py", line 181, in __exit__
  File "bash\gui\base_components.py", line 326, in destroy_component
RuntimeError: wrapped C/C++ object of type Dialog has been deleted